refactor(install): split helper modules
